refactor: refactor custome hook
